backend/requirements.txt
alembic==1.15.1
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
fastapi==0.115.11
greenlet==3.1.1
h11==0.14.0
idna==3.10
Mako==1.3.9
MarkupSafe==3.0.2
pydantic==2.10.6
pydantic_core==2.27.2
PyMySQL==1.1.1
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.46.0
typing_extensions==4.12.2
unicorn==2.1.3
uvicorn==0.34.0
